feat: fixes the last problem with the sanitising function
Refs: iris-gateway/iris-backlog#74
